Update Fractional_Knapsack
This algorithm is based on greedy approach and can be implemented in various language. It works in following way such as there is a knapsack whose capacity is chosen. After that the items are put into the sack once the capacity is fulfilled we stop. This algorithm doesn't guarantee maximum profit.